don't use readFile, store lua script in strings
